Fixed calcDirs()
Fixed a bug where the order that planets were created in affected how the simulation ran (the same planets but created in a different order yielded a different simulation). This issue was traced back to how calcDirs() handles points vertically in a line. The new simulation consistency can be seen by running config 1 vs 2.